### 2018-10-19 #1030 @whedon
[PRE REVIEW]: EMViz (Early Music Visualization): MATLAB runtime application
**Submitting author:** @carterenyi (<a href="http://orcid.org/0000-0003-2614-4777">Aaron Carter-Enyi</a>)
**Repository:** <a href="https://github.com/carterenyi/emviz" target ="_blank">https://github.com/carterenyi/emviz</a>
**Version:** v1.0.0
