### 2018-10-19 #1030 @arfon
Hi @carterenyi. Thanks for your submission to JOSS. A few things that need addressing before we can continue:

- Please format your paper metadata _exactly_ as you will find in [this example](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). YAML is whitespace sensitive and so your current metadata won't compile
- You seem to have content from our user guide in your `paper.md` file. Please remove this.
- There isn't a file called `paper.bib` so your bibliography won't compile into your paper
- There is no README or other documentation for the project
- There doesn't appear to be any source code in your project, instead there just appears to be an executable file
- There doesn't seem to be an open source license on your repository
