+++ paper.md
@@ -1,28 +1,28 @@
+---
 title: ‘EMViz (Early Music Visualization): MATLAB runtime application’
 tags:
-- MATLAB
-- music
-- arc diagrams
-- visualization
-- MIDI
+  - MATLAB
+  - music
+  - arc diagrams
+  - visualization
+  - MIDI
 authors:
-- name: Aaron Carter-Ényì
-orcid: 0000-0003-2614-4777
-affiliation: "1, 2, 3" # (Multiple affiliations must be quoted)
+  - name: Aaron Carter-Ényì
+    orcid: 0000-0003-2614-4777
+    affiliation: "1, 2, 3" # (Multiple affiliations must be quoted)
 affiliations:
-- name: Assistant Professor, Morehouse College
-index: 1
-- name: Instructor, Spelman College
-index: 2
-- name: Fellow, American Council of Learned Societies
-index: 3
+  - name: Assistant Professor, Morehouse College
+    index: 1
+  - name: Instructor, Spelman College
+    index: 2
+  - name: Fellow, American Council of Learned Societies
+    index: 3
 date: 16 October 2018
 bibliography: paper.bib
 ---
 
-250‑1000 words.
-
 #Summary
+
 EMViz (Early Music Visualization) provides built-in pattern recognition for symbolic music (MIDI) based on a contour recursion algorithm by Carter-Enyi [@Carter-Enyi:2016a].producing visualizations of musical form using arc diagrams, as proposed by Wattenberg ([@Wattenberg:2002].). Currently implemented in MATLAB and deployed as a standalone executable (using MATLAB Runtime), EMViz is now available at emviz.org and https://github.com/carterenyi/emviz (MIT license).
 
 A primary learning objective in music theory courses is the ability to analyze musical form, often using a bracketed diagram. This humanistic process often involves listening with a score, stopping the recording for a close reading of the score, flipping back and forth through the document to make comparisons, and so on. While manual analysis is important, the conventional method of summarizing one’s analysis as a bracket diagram is reductive without being illustrative. A promising digital method for visualization of musical form is found in Martin Wattenberg’s Shape of Song, a defunct web app from 2002. The arc diagram visualization technique for Shape of Song is brilliant, but ultimately the project did not live up to its potential because of a poor understanding of how composers develop musical themes. Wattenberg describes his string-matching method in his 2002 paper: “Each arc connects two matching passages, where a ‘match’ means that they contain the same sequence of pitches” [@Wattenberg:2002]. However, musical themes and motives are often transposed or otherwise modified in the course of a piece. A concept learned by music majors is the subject and answer of a fugue (an imitative polyphonic composition). The subject is the initial statement of a fugue’s theme, stated in a monophonic texture (by itself) and the answer is the first interpolation stated in a polyphonic texture (against counterpoint). Sometimes the answer is an exact transposition (real answer) and sometimes the answer is transposed and modified (tonal answer). In his paper, Wattenberg offers the possibility of matching melodic intervals (instead of pitch-strings), but this too would not match a subject to its tonal answer. So, how do we formalize the relationship between a subject and tonal answer? Or, the improvisational “licks” of Dizzy Gillespie? Such that we could define “themes” as a category that may be formalized computationally? Especially, when a computer does not know where to look for themes or what modifications to expect? Algorithms informed by humanistic understanding of music work much better than those that are not.
@@ -33,11 +33,10 @@ The computationally-efficient addition of the contour level data preparation (se
 
[…]
