+++ EMVizScripts/arcPlot.m
@@ -1,0 +1,291 @@
+function r=arcPlot(r,nmat,nmatFound,plotTitle,sortYN,rselect,stringNames)
+%% 1. Run findPattern first
+%% 2. arc of similarity
+%% if no time input is specified, it is animated
+%data needed:
+%nmat
+%nmatRCE
+%r (output of CT)
+%with ind or segind
+%and simind
+
+if nargin==1
+    nmat=r.nmat;
+    nmatFound=r.nmatFound;
+    r=r.r;
+    plotTitle='Arc Diagram of MIDI file';
+    sortYN=0;
+end
+
+%see if nargin = 6 below
+
+
+chans=unique(nmatFound(:,3));
+if numel(chans) > 1
+    nmatFound=nmatMono2poly(nmat,nmatFound);
+end
+
+%clear non-recursive strings
+for i=numel(r):-1:1
+    try
+        if numel(r(i).segind)==1 && numel(r(i).invind)==0
+            r(i)=[];
+        end
+    catch
+        if numel(r(i).ind)==1 && numel(r(i).simind)==0
+            r(i)=[];
+        end
+    end
+end
+
+if nargin == 7
+    r=r(rselect);
+end
+
+
+
+for i=1:numel(r)
+    clearvars ind2 starts2
+    try
+        segind=r(i).ind;
+        
+        for j=1:numel(segind)
+            starts2(j)=nmatFound(segind(j),1);
+        end
+        [starts2, ind2]=sort(starts2);
+        r(i).ind=segind(ind2);
+        
+    catch
+        segind=r(i).segind;
+        
+        for j=1:numel(segind)
+            %             j
+            %             segind(j)
+            %             size(nmatFound)
+            starts2(j)=nmatFound(segind(j),1);
+        end
+        [starts2, ind2]=sort(starts2);
+        r(i).segind=segind(ind2);
+    end
+    
+end
+
+if sortYN==1
+    for i=1:numel(r)
+        try
+            starts(i)=r(i).segind(1);
+        catch
+            starts(i)=r(i).ind(1);
+        end
+    end
+    [starts, ind]=sort(starts);
+    r=r(ind);
+end
+
+
+
+
+
+
+colornum=1;
+
+cmap1 = jet(numel(r));
+% color1=cmap1(2,:);
+% cmap1(2,:)=[];
+% cmap1=[color1;cmap1];
+
+
+screensize = get( groot, 'Screensize' );
+fig=figure();
+set(fig,'Color','w','Name','Video-EASE: arcPlot','Position', screensize*.9);
+
+midiplot(1)=plot([nmat(1,1) (nmat(1,1)+nmat(1,2))], [nmat(1,4) nmat(1,4)],'Color','k','LineWidth', 10);
+legendplots(1)=midiplot(1);
+hold on
+title(plotTitle,'FontSize',30);
+xlabel('Time (beats)','FontSize',16);
+ylabel('MIDI pitch (C4=60)','FontSize',16);
+set(gca,'FontSize',16,'FontWeight','bold');
+
+%initialize colors and labels for legend
+labels{1}='MIDI';
+for i=1:numel(r)
+    if nargin == 7 %%%%%
+        labels{end+1}=stringNames{i};
+    else
+        labels{end+1}=['String' sprintf(' %d',i)];
+    end
+    if numel(r)==1
+        h(i)=plot([-1,-2],[1,1],'Color',[0,0,1],'LineWidth',10); 
+    else
+    h(i)=plot([-1,-2],[1,1],'Color',cmap1(i,:),'LineWidth',10);
+    end
+    legendplots(end+1)=h(i);
+end
+
+
[…]
