+++ EMVizScripts/searchAndRank.m
@@ -1,0 +1,320 @@
+function [r,CL]=searchAndRank(nmat,type,cmin,in)
+
+%INPUTS
+%'nmat' is note matrix output from midi2nmat, this should be monophonic
+%with only one voice, if it is polyphonic poly2mono should be used first
+%'type' determines that type of search: '1' is pitches only; '2' is durations
+%only and '3' is pitches and durations
+%'cmin' is minimum cardinality for patterns, the algorithm will not search
+%for strings short than this number, a good minimum is five events
+%'in' is a structure for various search parameters see 'if nargin==4... below'
+
+%CAUTION:
+%1. this is only for monophonic pieces (if it is polyphonic run poly2mono
+%first)
+%2. does not work with human playback settings
+%(turn off in Finale or Sibelius before exporting a MIDI)
+
+%% Search Parameters from 'in'
+if nargin==4 % number of arguments into the function
+    try
+        degrees=in.degrees; % degrees determines how big the contour reduction window is
+    catch
+        degrees=2; % the standard is 2
+    end
+
+    try
+        cmax=in.cmax; % cardinality maximum sets the max string length
+    catch
+        cmax=50; % typically no strings longer than 50 are found anyway
+    end
+    
+
+    
+    try
+        usetime=in.usetime;
+    catch
+        usetime=0;
+    end
+
+    
+else %if there is no 'in' (i.e. nargin==3) these are default search parameters
+    degrees=2;
+    cmax=50;
+    usetime=0; %consider time when calculating coverage
+    %useDCL=0; % this is now separate input argument
+end
+
+
+
+
+rawnmat=nmat;
+
+
+pitches=nmat(:,4); 
+pitches=pitches';
+contcom=CT_makeContCom(pitches,degrees);
+CL=nansum(contcom);
+%CL=sum(contcom);
+
+
+durs=nmat(:,6);
+durs=durs(2:end)-durs(1:end-1);
+durs(end+1)=nmat(end,7);
+contcom=CT_makeContCom(durs,degrees);
+DCL=nansum(contcom);
+%DCL=sum(contcom);
+
+if type==1
+    CL=[CL;CL];
+elseif type ==2
+    CL=[DCL;DCL];
+elseif type ==3
+    CL=[CL;DCL];
+end
+
+for n=cmin:cmax
+    clearvars segments segmentu segmentucount segmentuind temp
+    for i=1:numel(pitches)-n+1
+        segments{i}=CL(:,i:i+n-1);
+        %         for j=1:degrees
+        %             try
+        %                 segments{i}(1:degrees+1-j,j)=-1;
+        %                 segments{i}(degrees+j:end,end+1-j)=-1;
+        %             catch
+        %             end
+        %         end
+    end
+    
+    
+    [segmentu, idx ,idx2] = uniquecell(segments);
+    %check to see if card saturation reached
+    if numel(segmentu)==numel(segments)
+        break
+    end
+    
+    
+    segmentucount=zeros(numel(segmentu),1);
+    for i=1:length(segmentu)
+        ind=find(idx2==i);
+        segmentuind{i}=ind;
+        segmentucount(i)=numel(ind);
+        %         numel(pitches)
+        %         segmentuind{i}(1)-2
+        %         segmentuind{i}(1)+n-1+2
+        try
+            segmentpitch{i}=pitches(segmentuind{i}(1)-2:segmentuind{i}(1)+n-1+2);
+        catch
+            try
[…]

[…]
