+++ EMVizScripts/searchAndRank_Annotated.m
@@ -1,0 +1,305 @@
+function [r,CL]=searchAndRank(nmat,type,cmin,in)
+
+%INPUTS
+%'nmat' is note matrix output from midi2nmat, this should be monophonic
+%with only one voice, if it is polyphonic poly2mono should be used first
+%'type' determines that type of search: '1' is pitches only; '2' is durations
+%only and '3' is pitches and durations
+%'cmin' is minimum cardinality for patterns, the algorithm will not search
+%for strings short than this number, a good minimum is five events
+%'in' is a structure for various search parameters see 'if nargin==4... below'
+
+%CAUTION:
+%1. this is only for monophonic pieces (if it is polyphonic run poly2mono
+%first)
+%2. does not work with human playback settings
+%(turn off in Finale or Sibelius before exporting a MIDI)
+
+%% Search Parameters from 'in'
+if nargin==4 % number of arguments into the function
+    
+    % other search parameters delgap, simmin, etc should be added here
+    
+    try
+        degrees=in.degrees; % degrees determines how big the contour reduction window is
+    catch
+        degrees=2; % the standard is 2
+    end
+
+    try
+        cmax=in.cmax; % cardinality maximum sets the max string length
+    catch
+        cmax=50; % typically no strings longer than 50 are found anyway
+    end
+    
+else %if there is no 'in' (i.e. nargin==3) these are default search parameters
+    degrees=2;
+    cmax=50;
+    %insert defaults for other search parameters
+end
+
+rawnmat=nmat; %store the unaltered nmat in 'rawnmat'
+
+%% Create Contour Level ('CL') Array to search
+pitches=nmat(:,4); %create an array of pitches
+pitches=pitches'; %transpose from vertical to horizontal
+contcom=CT_makeContCom(pitches,degrees); % create a continuous contour matrix
+CL=nansum(contcom); % sum the columns of the matrix into contour levels, ignoring nans
+
+%% Create Durational Contour Level ('DCL') Array to search
+onsets=nmat(:,6); %create an array of onsets in seconds
+durs=onsets(2:end)-onsets(1:end-1); %convert to interonset durations by subtracting n-1 from n;
+durs(end+1)=nmat(end,7); %add the last duration to the end so that there are values for all events
+contcom=CT_makeContCom(durs,degrees); % create a continuous contour matrix
+DCL=nansum(contcom); % sum the columns of the matrix into contour levels, ignoring nans
+
+%% create a searchable matrix (2 x number of events) based on type input argument
+if type==1
+    CL=[CL;CL]; %only pitch information
+elseif type==2
+    CL=[DCL;DCL]; %only durational information
+elseif type==3
+    CL=[CL;DCL]; %combining pitch and durational information
+end
+
+%% Search the matrix
+%create a segment pool and find repeating segments (strings)
+for n=cmin:cmax % for loop
+    clearvars segments segmentu segmentucount segmentuind temp
+    
+    for i=1:numel(pitches)-n+1
+        segments{i}=CL(:,i:i+n-1);
+    end
+
+    [segmentu, idx ,idx2] = uniquecell(segments);
+    
+    %check to see if card saturation reached
+    if numel(segmentu)==numel(segments)
+        break
+    end
+    
+    
+    segmentucount=zeros(numel(segmentu),1);
[…]
