+++ EMVizScripts/EMVizGUI.m
@@ -1,0 +1,446 @@
+function varargout = EMVizGUI(varargin)
+% EMVIZGUI MATLAB code for EMVizGUI.fig
+%      EMVIZGUI, by itself, creates a new EMVIZGUI or raises the existing
+%      singleton*.
+%
+%      H = EMVIZGUI returns the handle to a new EMVIZGUI or the handle to
+%      the existing singleton*.
+%
+%      EMVIZGUI('CALLBACK',hObject,eventData,handles,...) calls the local
+%      function named CALLBACK in EMVIZGUI.M with the given input arguments.
+%
+%      EMVIZGUI('Property','Value',...) creates a new EMVIZGUI or raises the
+%      existing singleton*.  Starting from the left, property value pairs are
+%      applied to the GUI before EMVizGUI_OpeningFcn gets called.  An
+%      unrecognized property name or invalid value makes property application
+%      stop.  All inputs are passed to EMVizGUI_OpeningFcn via varargin.
+%
+%      *See GUI Options on GUIDE's Tools menu.  Choose "GUI allows only one
+%      instance to run (singleton)".
+%
+% See also: GUIDE, GUIDATA, GUIHANDLES
+
+% Edit the above text to modify the response to help EMVizGUI
+
+% Last Modified by GUIDE v2.5 28-Oct-2017 09:19:53
+
+% Begin initialization code - DO NOT EDIT
+gui_Singleton = 1;
+gui_State = struct('gui_Name',       mfilename, ...
+                   'gui_Singleton',  gui_Singleton, ...
+                   'gui_OpeningFcn', @EMVizGUI_OpeningFcn, ...
+                   'gui_OutputFcn',  @EMVizGUI_OutputFcn, ...
+                   'gui_LayoutFcn',  [] , ...
+                   'gui_Callback',   []);
+if nargin && ischar(varargin{1})
+    gui_State.gui_Callback = str2func(varargin{1});
+end
+
+if nargout
+    [varargout{1:nargout}] = gui_mainfcn(gui_State, varargin{:});
+else
+    gui_mainfcn(gui_State, varargin{:});
+end
+% End initialization code - DO NOT EDIT
+
+
+% --- Executes just before EMVizGUI is made visible.
+function EMVizGUI_OpeningFcn(hObject, eventdata, handles, varargin)
+% This function has no output args, see OutputFcn.
+% hObject    handle to figure
+% eventdata  reserved - to be defined in a future version of MATLAB
+% handles    structure with handles and user data (see GUIDATA)
+% varargin   command line arguments to EMVizGUI (see VARARGIN)
+
+% Choose default command line output for EMVizGUI
+handles.output = hObject;
+handles.filename=[];
+handles.pathname=[];
+handles.stringSelect{1}='initialize';
+handles.stringNumbers=[];
+handles.r=[];
+handles.nmatFound=[];
+handles.nmat=[];
+xlabel(handles.axes1,'Time (Beats)');
+ylabel(handles.axes1,'Pitch (C4=60)');
+ylim([55 67])
+xlim([0 8])
+
+
+% Update handles structure
+guidata(hObject, handles);
+
+% UIWAIT makes EMVizGUI wait for user response (see UIRESUME)
+% uiwait(handles.figure1);
+
+
+% --- Outputs from this function are returned to the command line.
+function varargout = EMVizGUI_OutputFcn(hObject, eventdata, handles) 
+% varargout  cell array for returning output args (see VARARGOUT);
+% hObject    handle to figure
+% eventdata  reserved - to be defined in a future version of MATLAB
+% handles    structure with handles and user data (see GUIDATA)
+
+% Get default command line output from handles structure
+varargout{1} = handles.output;
+
+
[…]
