+++ README.md
@@ -1,1 +1,3 @@
 # emviz
+Please see "EMViz_Manual.pdf" for instructions on how to download, install and use the standalone application from www.emviz.org
+Source code is available in the "EMVizScripts" folders 
