### 2018-10-23 c3ab4c9
Update paper.md
