### 2018-10-23 0b4a7a9
Update paper.md
