### 2018-10-23 804a636
Delete paper.bib
