+++ paper.bib
@@ -1,83 +1,0 @@
-﻿@article{RN2925,
-   author = {Carter-Enyi, Aaron},
-   title = {Contour Recursion and Auto-Segmentation},
-   journal = {Music Theory Online},
-   volume = {22},
-   number = {1},
-   year = {2016},
-   type = {Journal Article}
-}
-
-@phdthesis{RN2929,
-   author = {Carter-Enyi, Aaron},
-   title = {Contour Levels: An Abstraction of Pitch Space based on African Tone Systems},
-   university = {Ohio State University},
-   year = {2016},
-   type = {Thesis}
-}
-
-@article{RN2905,
-   author = {Carter-Enyi, Aaron},
-   title = {Hooked on Sol-Fa: The Do-Re-Mi Heuristic for Yoruba Speech Tones},
-   journal = {Africa},
-   volume = {88},
-   number = {2},
-   pages = {267-290},
-   year = {2018},
-   type = {Journal Article}
-}
-
-@book{RN160,
-   author = {Huron, David},
-   title = {Sweet Anticipation: Music and the Psychology of Expectation},
-   publisher = {MIT Press},
-   address = {Cambridge},
-   year = {2006},
-   type = {Book}
-}
-
-@book{RN1833,
-   author = {Morris, Robert D},
-   title = {Composition with pitch-classes: a theory of compositional design},
-   publisher = {Yale University Press},
-   ISBN = {0300036841},
-   year = {1987},
-   type = {Book}
-}
-
-@article{RN1612,
-   author = {Quinn, Ian},
-   title = {Fuzzy Extensions to the Theory of Contour},
-   journal = {Music Theory Spectrum},
-   volume = {19},
-   number = {2},
-   pages = {232-263},
-   ISSN = {01956167},
-   DOI = {10.2307/745755},
-   url = {http://www.jstor.org.proxy.lib.ohio-state.edu/stable/745755},
-   year = {1997},
-   type = {Journal Article}
-}
-
-@article{RN3100,
-   author = {Thomassen, Joseph M},
-   title = {Melodic accent: Experiments and a tentative model},
-   journal = {The Journal of the Acoustical Society of America},
-   volume = {71},
-   number = {6},
-   pages = {1596-1605},
-   ISSN = {0001-4966},
-   year = {1982},
-   type = {Journal Article}
-}
-
-@inproceedings{RN2913,
-   author = {Wattenberg, Martin},
-   title = {Arc diagrams: Visualizing structure in strings},
-   booktitle = {Information Visualization, 2002. INFOVIS 2002. IEEE Symposium on},
-   publisher = {IEEE},
-   pages = {110-116},
-   ISBN = {076951751X},
-   type = {Conference Proceedings}
-}
-
