### 2018-10-23 6af9fd4
Create paper.bib
