### 2018-10-23 c804810
Update paper.bib
