+++ paper.bib
@@ -1,4 +1,4 @@
-@article{RN2925,
+@article{Carter-Enyi:2016,
    author = {Carter-Enyi, Aaron},
    title = {Contour Recursion and Auto-Segmentation},
    journal = {Music Theory Online},
@@ -27,7 +27,7 @@
    type = {Journal Article}
 }
 
-@book{RN160,
+@book{Huron:2006,
    author = {Huron, David},
    title = {Sweet Anticipation: Music and the Psychology of Expectation},
    publisher = {MIT Press},
@@ -36,7 +36,7 @@
    type = {Book}
 }
 
-@book{RN1833,
+@book{Morris:1987,
    author = {Morris, Robert D},
    title = {Composition with pitch-classes: a theory of compositional design},
    publisher = {Yale University Press},
@@ -45,7 +45,7 @@
    type = {Book}
 }
 
-@article{RN1612,
+@article{Quinn:1997,
    author = {Quinn, Ian},
    title = {Fuzzy Extensions to the Theory of Contour},
    journal = {Music Theory Spectrum},
@@ -59,7 +59,7 @@
    type = {Journal Article}
 }
 
-@article{RN3100,
+@article{Thomassen:1982,
    author = {Thomassen, Joseph M},
    title = {Melodic accent: Experiments and a tentative model},
    journal = {The Journal of the Acoustical Society of America},
@@ -71,7 +71,7 @@
    type = {Journal Article}
 }
 
-@inproceedings{RN2913,
+@inproceedings{Wattenberg:2002,
    author = {Wattenberg, Martin},
    title = {Arc diagrams: Visualizing structure in strings},
    booktitle = {Information Visualization, 2002. INFOVIS 2002. IEEE Symposium on},
