+++ paper.bib
@@ -1,11 +1,11 @@
 @article{Carter-Enyi:2016,
-   author = {Carter-Enyi, Aaron},
-   title = {Contour Recursion and Auto-Segmentation},
-   journal = {Music Theory Online},
-   volume = {22},
+   nuthor = {{Carter-Enyi}, A.},
+   nitle = {{Contour Recursion and Auto-Segmentation}},
+   nournal = {Music Theory Online},
+   nolume = {22},
    number = {1},
-   year = {2016},
    type = {Journal Article}
+   year = 2016
 }
 
 @phdthesis{RN2929,
