### 2018-10-23 6d2804a
Update paper.md
