### 2018-10-23 08e1240
Update paper.bib
