+++ paper.bib
@@ -1,3 +1,68 @@
+@article{Pearson:2017,
+    Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+    Adsurl = {http://adsabs.harvard.edu/abs/2017arXiv170304627P},
+    Archiveprefix = {arXiv},
+    Author = {{Pearson}, S. and {Price-Whelan}, A.~M. and {Johnston}, K.~V.},
+    Eprint = {1703.04627},
+    Journal = {ArXiv e-prints},
+    Keywords = {Astrophysics - Astrophysics of Galaxies},
+    Month = mar,
+    Title = {{Gaps in Globular Cluster Streams: Pal 5 and the Galactic Bar}},
+    Year = 2017
+}
+
+@book{Binney:2008,
+    Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+    Adsurl = {http://adsabs.harvard.edu/abs/2008gady.book.....B},
+    Author = {{Binney}, J. and {Tremaine}, S.},
+    Booktitle = {Galactic Dynamics: Second Edition, by James Binney and Scott Tremaine.~ISBN 978-0-691-13026-2 (HB).~Published by Princeton University Press, Princeton, NJ USA, 2008.},
+    Publisher = {Princeton University Press},
+    Title = {{Galactic Dynamics: Second Edition}},
+    Year = 2008
+}
+
+@article{zenodo,
+    Abstractnote = {Gala is a Python package for Galactic astronomy and gravitational dynamics. The bulk of the package centers around implementations of gravitational potentials, numerical integration, and nonlinear dynamics.},
+    Author = {Adrian Price-Whelan and Brigitta Sipocz and Syrtis Major and Semyeong Oh},
+    Date-Modified = {2017-08-13 14:14:18 +0000},
+    Doi = {10.5281/zenodo.833339},
+    Month = {Jul},
+    Publisher = {Zenodo},
+    Title = {adrn/gala: v0.2.1},
+    Year = {2017},
+    Bdsk-Url-1 = {http://dx.doi.org/10.5281/zenodo.833339}
+}
+
+@article{gaia,
+    author = {{Gaia Collaboration}},
+    title = "{The Gaia mission}",
+    journal = {\aap},
+    archivePrefix = "arXiv",
+    eprint = {1609.04153},
+    primaryClass = "astro-ph.IM",
+    keywords = {space vehicles: instruments, Galaxy: structure, astrometry, parallaxes, proper motions, telescopes},
+    year = 2016,
+    month = nov,
+    volume = 595,
+    doi = {10.1051/0004-6361/201629272},
+    adsurl = {http://adsabs.harvard.edu/abs/2016A%26A...595A...1G},
+}
+
+@article{astropy,
+    author = {{Astropy Collaboration}},
+    title = "{Astropy: A community Python package for astronomy}",
+    journal = {\aap},
+    archivePrefix = "arXiv",
+    eprint = {1307.6212},
+    primaryClass = "astro-ph.IM",
+    keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
+    year = 2013,
+    month = oct,
+    volume = 558,
+    doi = {10.1051/0004-6361/201322068},
+    adsurl = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A}
+}
+
 @article{Carter-Enyi:2016,
    nuthor = {{Carter-Enyi}, A.},
    nitle = {{Contour Recursion and Auto-Segmentation}},
