### 2018-10-23 b6466af
Update paper.md
