### 2018-10-23 cbfa03d
Update paper.bib
