+++ paper.bib
@@ -10,7 +10,7 @@
 
 @phdthesis{Carter-Enyi:2016b,
    author = {Carter-Enyi, Aaron},
-   title = {Contour Levels: An Abstraction of Pitch Space based on African Tone Systems},
+   title = {{Contour Levels: An Abstraction of Pitch Space based on African Tone Systems}},
    university = {Ohio State University},
    type = {Thesis},
    year = 2016
@@ -18,7 +18,7 @@
 
 @book{Huron:2006,
    author = {Huron, David},
-   title = {Sweet Anticipation: Music and the Psychology of Expectation},
+   title = {{Sweet Anticipation: Music and the Psychology of Expectation}},
    publisher = {MIT Press},
    address = {Cambridge},
    type = {Book},
@@ -27,7 +27,7 @@
 
 @book{Morris:1987,
    author = {Morris, Robert D},
-   title = {Composition with pitch-classes: a theory of compositional design},
+   title = {{Composition with pitch-classes: a theory of compositional design}},
    publisher = {Yale University Press},
    ISBN = {0300036841},
    type = {Book},
@@ -36,7 +36,7 @@
 
 @article{Quinn:1997,
    author = {Quinn, Ian},
-   title = {Fuzzy Extensions to the Theory of Contour},
+   title = {{Fuzzy Extensions to the Theory of Contour}},
    journal = {Music Theory Spectrum},
    volume = {19},
    number = {2},
@@ -50,7 +50,7 @@
 
 @article{Thomassen:1982,
    author = {Thomassen, Joseph M},
-   title = {Melodic accent: Experiments and a tentative model},
+   title = {{Melodic accent: Experiments and a tentative model}},
    journal = {The Journal of the Acoustical Society of America},
    volume = {71},
    number = {6},
@@ -62,7 +62,7 @@
 
 @inproceedings{Wattenberg:2002,
    author = {Wattenberg, Martin},
-   title = {Arc diagrams: Visualizing structure in strings},
+   title = {{Arc diagrams: Visualizing structure in strings}},
    booktitle = {Information Visualization, 2002. INFOVIS 2002. IEEE Symposium on},
    publisher = {IEEE},
    pages = {110-116},
