+++ paper.md
@@ -34,7 +34,7 @@ There are three primary components of the EMViz MATLAB Runtime application: (1) 
 The computationally-efficient addition of the contour level data preparation (see Carter-Enyi 2016b) [@Carter-Enyi:2016b] makes equivalence-based pattern recognition (e.g. string-matching) much more effective. Notably, this pre-processing of the symbolic data makes significant improvements to performance with negligible added runtime. ATAVizM’s algorithm identifies 100% of the subjects and approximately 80% of all statements in the 48 Fugues of Bach’s Well-Tempered Clavier.
 
 
-The algorithm brings together contour theory ([@Morris:1987]., [@Quinn:1997].) with studies of melodic accent ([@Thomassen:1982]., [@Huron:2006].). Symbolic music data (.midi, .xml) from various sources (including ELVIS at McGill and the Yale Classical Archives Corpus) may be imported, analyzed and visualized in a matter of minutes. The contour algorithm behind EMViz uses local contour comparisons within a 2-degree radius of each focus pitch. This makes use of Quinn’s binary C+ comparison, where 0 is equal or below and 1 is above the reference pitch [@Quinn:1997]. In Figure 1, the subject and tonal answer of Bach’s C-minor Fugue (BWV 847) are shown to be equivalent strings using this method (they both produce the same “continuous” matrix).
+The algorithm brings together contour theory ([@Morris:1987], [@Quinn:1997]) with studies of melodic accent ([@Thomassen:1982], [@Huron:2006]). Symbolic music data (.midi, .xml) from various sources (including ELVIS at McGill and the Yale Classical Archives Corpus) may be imported, analyzed and visualized in a matter of minutes. The contour algorithm behind EMViz uses local contour comparisons within a 2-degree radius of each focus pitch. This makes use of Quinn’s binary C+ comparison, where 0 is equal or below and 1 is above the reference pitch [@Quinn:1997]. In Figure 1, the subject and tonal answer of Bach’s C-minor Fugue (BWV 847) are shown to be equivalent strings using this method (they both produce the same “continuous” matrix).
 
 ![Example figure1.png](figure1.png)
 
