### 2018-10-23 #1030 @carterenyi
Done, thank you.

On Tue, Oct 23, 2018 at 11:03 AM Arfon Smith <notifications@github.com>
wrote:

> @carterenyi <https://github.com/carterenyi> - please merge this PR which
> will help your paper compile:
> https://github.com/carterenyi/emviz/pull/2/files
>
> In addition:
>
> *Please update your paper.bib file as your citations currently can't be
> found*
>
> pandoc-citeproc: reference Carter-Enyi:2016a not found
> pandoc-citeproc: reference Wattenberg:2002 not found
> pandoc-citeproc: reference Wattenberg:2002 not found
> pandoc-citeproc: reference Morris:1987 not found
> pandoc-citeproc: reference Quinn:1997 not found
> pandoc-citeproc: reference Thomassen:1982 not found
> pandoc-citeproc: reference Huron:2006 not found
> pandoc-citeproc: reference Quinn:1997 not found
>
> *Please update the figure references*
>
> These links to the figures don't link to anything and need updating:
>
> ![Example figure1.] Example arc diagram visualizations include music from the Liber
> Usualis ![Example figure2.], Josquin des Prez ![Example figure3.]  and J. S. Bach
> ![Example figure4.].
>
