### 2018-10-23 #1030 @arfon
Looks like you're missing a comma (`,`) on this line: https://github.com/carterenyi/emviz/blob/master/paper.bib#L7

@carterenyi - please make this change and then ask Whedon to compile the paper with `@whedon generate pdf`
