+++ paper.bib
@@ -4,7 +4,7 @@
    nournal = {Music Theory Online},
    nolume = {22},
    number = {1},
-   type = {Journal Article}
+   type = {Journal Article},
    year = 2016
 }
 
