### 2018-10-23 #1030 @carterenyi
On Tue, Oct 23, 2018 at 3:21 PM Arfon Smith <notifications@github.com>
wrote:

> Looks like you're missing a comma (,) on this line:
> https://github.com/carterenyi/emviz/blob/master/paper.bib#L7
>
> @carterenyi <https://github.com/carterenyi> - please make this change and
> then ask Whedon to compile the paper with @whedon generate pdf
>
