### 2018-10-24 #1030 @carterenyi
Article proof checked, do I need to do anything else for it to move on to review?
