### 2018-10-25 #1030 @arfon
> Article proof checked, do I need to do anything else for it to move on to review?

Next we need to find a handling editor for this submission.

@Kevin-Mattheus-Moerman @cMadan - would you be willing to edit this submission for JOSS?
