### 2018-10-27 #1030 @arfon
> @whedon assign @Kevin-Mattheus-Moerman as editor

Thanks for taking this on @Kevin-Mattheus-Moerman ❤️
