### 2018-10-30 #1030 @Kevin-Mattheus-Moerman
@carterenyi Before we proceed can you please work on the following points:

- [ ] Can you please expand your README.md. It is the landing page for your project and should inform potential users/contributors about what the project is, how to get started, how to contribute, what the license is etc. Here is a basic template: https://gist.github.com/PurpleBooth/109311bb0361f32d87a2
Here is an example of one of mine: https://github.com/gibbonCode/GIBBON/blob/master/README.md

- [ ] Can you please add community/contributing guidelines (e.g. CONTRIBUTING.md) and a code of conduct document (e.g. COC.md you can find a good standard template here: https://www.contributor-covenant.org/). I recommend you link to the contributing guidelines in the readme and to the code of conduct in the contributing guidelines.

See also these resources:
https://help.github.com/articles/setting-guidelines-for-repository-contributors/
https://opensource.guide/code-of-conduct/

Examples:
https://github.com/atom/atom/blob/master/CONTRIBUTING.md
https://github.com/rails/rails/blob/master/CONTRIBUTING.md
