### 2018-11-10 399e622
Create contributing.md
