+++ contributing.md
@@ -1,0 +1,23 @@
+## How to contribute to emviz: Early Music Visualization
+
+#### **Did you find a bug?**
+
+Please email <carterenyi@gmail.com>
+
+#### **Did you write a patch that fixes a bug?**
+
+* Open a new GitHub pull request with the patch.
+
+* Ensure the PR description clearly describes the problem and solution.
+
+#### **Do you intend to add a new feature or change an existing one?**
+
+Please email <carterenyi@gmail.com>
+
+#### **Do you have questions about the source code?**
+
+Please email <carterenyi@gmail.com>
+
+#### **Do you want to contribute to the emviz documentation?**
+
+Please email <carterenyi@gmail.com>
