### 2018-11-10 #1030 @carterenyi
I've greatly expanded the readme and added the contributing.md and coc.md.

if @ssfrr is not able to review it, are there other reviewers available?

On Tue, Oct 30, 2018 at 8:12 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @ssfrr <https://github.com/ssfrr> is this your 🍵 (i.e. your cup of tea)?
> Would you be able to review this work?
>
