### 2018-11-17 #1030 @carterenyi
Hello,

Any progress on this?

On Sat, Nov 10, 2018 at 11:16 AM Aaron Carter-Enyi <carterenyi@gmail.com>
wrote:

> I've greatly expanded the readme and added the contributing.md and coc.md.
>
> if @ssfrr is not able to review it, are there other reviewers available?
>
> On Tue, Oct 30, 2018 at 8:12 AM Kevin Mattheus Moerman <
> notifications@github.com> wrote:
>
>> @ssfrr <https://github.com/ssfrr> is this your 🍵 (i.e. your cup of
>> tea)? Would you be able to review this work?
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/1030#issuecomment-434276391>,
>> or mute the thread
>> <https://github.com/notifications/unsubscribe-auth/AXim6ipuggf2syAyOLKLwz_3b__dLUyCks5uqEIVgaJpZM4Xvdx1>
>> .
>>
>
>
> --
> Aaron Carter-Ényì, PhD, CAGO
>

--
Aaron Carter-Ényì, PhD, CAGO
