### 2018-11-20 #1030 @Kevin-Mattheus-Moerman
@alexanderlerch @tyiannak @nikhilbhanu @ybayle @gwynethbradbury @hooman650 @aparks5 this is a review invitation for [JOSS (the Journal of Open Source Software)](http://joss.theoj.org/). Would you be able to review [this work](https://github.com/carterenyi/emviz), it is a MATLAB project featuring "pattern recognition for symbolic music".
The [JOSS review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) focuses largely on [the software](https://github.com/carterenyi/emviz) and a very [short paper](https://github.com/openjournals/joss-papers/blob/joss.01030/joss.01030/10.21105.joss.01030.pdf).

Please let me know if you are able to review this work, and let me know if you have any questions.
If your expertise does not fully overlap with this work but you are still able to review several aspects of it we'd still welcome your comments. Thank you!

If you agree to review you will be assigned as a reviewer and I will open a dedicated review issue where the actual review process takes place.
