### 2018-11-20 #1030 @ybayle
Hello @Kevin-Mattheus-Moerman and @carterenyi and thank you for your invitation.

I agree to review this submission.
If you are missing reviewers, you could ask people from the ISMIR community (On [Slack](https://launchpass.com/mircommunity) or on the [Google Mailing List](https://groups.google.com/a/ismir.net/forum/#!forum/community).

Little details on the form of the last proof-version before starting the review on the content:
- The 7th reference is badly displayed,
- Images' titles only contain the file name and no description,
- There is a missing `addpath('../EMVizScripts/');` in `EMVizGUI.m` in order to launch the project after a git clone.
