+++ paper.bib
@@ -1,8 +1,8 @@
 @article{Carter-Enyi:2016a,
-   nuthor = {{Carter-Enyi}, A.},
-   nitle = {{Contour Recursion and Auto-Segmentation}},
-   nournal = {Music Theory Online},
-   nolume = {22},
+   author = {{Carter-Enyi}, Aaron},
+   title = {{Contour Recursion and Auto-Segmentation}},
+   journal = {Music Theory Online},
+   volume = {22},
    number = {1},
    type = {Journal Article},
    year = 2016
