### 2018-11-21 e05735c
Update paper.md
