+++ paper.md
@@ -29,28 +29,22 @@ A primary learning objective in music theory courses is the ability to analyze m
 
 There are three primary components of the EMViz MATLAB Runtime application: (1) Data Preparation (Pre-Processing); (2) Pattern Recognition; and (3) Visualization. These components are compared between Shape of Song, EMViz (now available in MATLAB Runtime, and ATAVizM (currently being developing Python) in Table 1.
 
-![Example table1.png](table1.png)
+![Table of features of Wattenberg's Shape of Song, EMViz (released 2017) and ATAVizM (in development)](table1.png)
 
 The computationally-efficient addition of the contour level data preparation (see Carter-Enyi 2016b) [@Carter-Enyi:2016b] makes equivalence-based pattern recognition (e.g. string-matching) much more effective. Notably, this pre-processing of the symbolic data makes significant improvements to performance with negligible added runtime. ATAVizM’s algorithm identifies 100% of the subjects and approximately 80% of all statements in the 48 Fugues of Bach’s Well-Tempered Clavier.
 
 
 The algorithm brings together contour theory ([@Morris:1987], [@Quinn:1997]) with studies of melodic accent ([@Thomassen:1982], [@Huron:2006]). Symbolic music data (.midi, .xml) from various sources (including ELVIS at McGill and the Yale Classical Archives Corpus) may be imported, analyzed and visualized in a matter of minutes. The contour algorithm behind EMViz uses local contour comparisons within a 2-degree radius of each focus pitch. This makes use of Quinn’s binary C+ comparison, where 0 is equal or below and 1 is above the reference pitch [@Quinn:1997]. In Figure 1, the subject and tonal answer of Bach’s C-minor Fugue (BWV 847) are shown to be equivalent strings using this method (they both produce the same “continuous” matrix).
 
-![Example figure1.png](figure1.png)
+![Contour level equivalence between the subject and tonal answer of Bach's Fugue in C minor BWV 847 based on Carter-Enyi 2016b [@Carter-Enyi:2016b]](figure1.png)
 
 Example arc diagram visualizations include music from the Liber Usualis, Josquin des Prez and J. S. Bach.
 
-Figure 2: Liber Usualis
+![EMViz analysis and visualization of "Alleluia, Exsultate Deo" chant from the Liber Usualis (11th century)](figure2.png)
 
-![Example figure2.png](figure2.png)
+![EMViz analysis and visualization of "Ave Maria" by Josquin des Prez (15-16th century)](figure3.png)
 
-Figure 3: Josquin des Prez
-
-![Example figure3.png](figure3.png)
-
-Figure 4: J. S. Bach
-
-![Example figure4.png](figure4.png)
+![EMViz analysis and visualization of Back's Fugue in C minor BWV 847 (18th century) (NB: an octave displacement is removed from one of the counter-subject 2 statements](figure4.png)
 
 In summary, EMViz (Early Music Visualization) identifies and implements three major improvements over Wattenberg’s Shape of Song: (1) a pattern-matching algorithm based on heuristics from music theory, (2) theme identification by user input or selection integrated into the application, and (3) color-coding of arcs (between matched patterns) and a legend to make the diagram more comprehensible, interpretive and utilitarian for research and teaching—instead of just a pretty picture.
 
