### 2018-11-21 #1030 @carterenyi
On Wed, Nov 21, 2018 at 1:43 AM Aaron Carter-Enyi <carterenyi@gmail.com>
wrote:

> @whedon <https://github.com/whedon> generate pdf
>
> On Tue, Nov 20, 2018 at 4:40 PM whedon <notifications@github.com> wrote:
>
>> OK, the reviewer is @ybayle <https://github.com/ybayle>
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/1030#issuecomment-440439225>,
>> or mute the thread
>> <https://github.com/notifications/unsubscribe-auth/AXim6h-9z1W49bltfIs6V8xHo7IABjPYks5uxHbfgaJpZM4Xvdx1>
>> .
>>
>
>
> --
> Aaron Carter-Ényì, PhD, CAGO
>

--
Aaron Carter-Ényì, PhD, CAGO
