### 2018-11-21 74a60b6
Update EMVizGUI.m
