### 2018-11-21 #1030 @carterenyi
@Kevin-Mattheus-Moerman and @ybayle
I have responded to Yann's noted issues:

   - The 7th reference is badly displayed,
   - Images' titles only contain the file name and no description,
   - There is a missing addpath('../EMVizScripts/'); in EMVizGUI.m in order
   to launch the project after a git clone.

I added a few lines for the last item, but I couldn't really tell whether
it is working because when I download a clone and run EMVizGUI.m, MATLAB
asks me if I want to add it to the path--which may be the red flag itself.
If it's not working as it should, please provide a little more instruction
if possible (like where and what to insert) -- I know signal processing but
very little about deployment.
