### 2018-11-21 #1030 @gwynethbradbury
Kevin,
I’d be happy to review the software and have some time over the next few days to do so. Let me know exactly how you wish for me to proceed.
Thanks,
Gwyneth

> On 20 Nov 2018, at 09:53, Kevin Mattheus Moerman <notifications@github.com> wrote:
>
> @alexanderlerch <https://github.com/alexanderlerch> @tyiannak <https://github.com/tyiannak> @nikhilbhanu <https://github.com/nikhilbhanu> @ybayle <https://github.com/ybayle> @gwynethbradbury <https://github.com/gwynethbradbury> @hooman650 <https://github.com/hooman650> @aparks5 <https://github.com/aparks5> this is a review invitation for JOSS (the Journal of Open Source Software) <http://joss.theoj.org/>. Would you be able to review this work <https://github.com/carterenyi/emviz>, it is a MATLAB project featuring "pattern recognition for symbolic music".
> The JOSS review process <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html> focuses largely on the software <https://github.com/carterenyi/emviz> and a very short paper <https://github.com/openjournals/joss-papers/blob/joss.01030/joss.01030/10.21105.joss.01030.pdf>.
>
> Please let me know if you are able to review this work, and let me know if you have any questions.
> If your expertise does not fully overlap with this work but you are still able to review several aspects of it we'd still welcome your comments. Thank you!
>
> If you agree to review you will be assigned as a reviewer and I will open a dedicated review issue where the actual review process takes place.
>
