### 2018-11-21 #1094 @whedon
[REVIEW]: EMViz (Early Music Visualization): MATLAB runtime application
**Submitting author:** @carterenyi (<a href="http://orcid.org/0000-0003-2614-4777">Aaron Carter-Enyi</a>)
**Repository:** <a href="https://github.com/carterenyi/emviz" target ="_blank">https://github.com/carterenyi/emviz</a>
**Version:** v1.0.0

- [x] **Authorship:**
  - [x] Has the submitting author (@carterenyi) made major contributions to the software?
  - [x] Does the full list of paper authors seem appropriate and complete?

  - [x] Works on Windows 7
