### 2018-11-21 #1094 @Kevin-Mattheus-Moerman
@ybayle, @gwynethbradbury, thanks for agreeing to review this submission for JOSS! :tada:
This is where the review process takes place. There are instructions and check boxes at the top of this issue that guide you through the process. Feel free to open issues you have on [the project repository](https://github.com/carterenyi/emviz), and please refer to them here as well e.g. by making a checkbox list (using `- [ ] `):

> Minor comment
> - [ ] I have this minor issue [LINK TO ISSUE 1]
>
> Major issues:
>
> - [ ] and this one [LINK TO ISSUE 2]

Let me know if you have any questions.
