### 2018-12-02 #1094 @ybayle
*[First part of the review]*

 `- Version: Does the release version given match the GitHub release (v1.0.0)?`

There is no release currently (cf https://github.com/carterenyi/emviz/releases). Would it be possible to have a released version or is it preferable to wait for the end of the review process?

`- Authorship: Does the full list of paper authors seem appropriate and complete?`

There is no clear context about the work done. There are no references to `William S. Dula` that is mentioned on the pdf on the windows download link nor to `Khalif Ali` that is mentioned as the proprietary of the repository where the installer and pdf are stored. Would it be possible to have more details about them as they are not mentioned in the list of authors?

`- Installation instructions.`

The installation procedure works fine on Windows 7. I haven't tested it yet on Mac nor Windows 10 but I will soon. For a direct git clone, there was a missing Matlab `addpath` that seems to have been resolved with the latest commit https://github.com/carterenyi/emviz/commit/74a60b6d49042d4cf0618d97599a7f70ba78c52d
I have to retry from scratch the install process to double check this.
Yet, there is still a problem with the program when cloned and launch from scratch. The GUI can handle the loading of a MIDI file and plot correctly but after the first plot, there are multiple error messages and no buttons react at all. This is a behaviour that I did not observe with the windows installer. Do you have the same problem on your side? Could you provide a detailed procedure to launch the GUI from a `git clone` in a Matlab environment?
