### 2018-12-11 #1094 @carterenyi
Response to @ybayle <https://github.com/ybayle>:

First of all, thank you!

Version:
I will work on that, if @Kevin-Mattheus-Moerman
<https://github.com/Kevin-Mattheus-Moerman> has any insight on the standard
for JOSS, it would be appreciated
Authorship:
I am the sole author of the MATLAB code. Mr. Dula and Mr. Ali were
undergrad students hired to translate the existing MATLAB code into Python.
Mr. Ali dropped from the project without completing any translations, Mr.
Dula is still working with me on the Python translation and when that is
released will be credited as a coauthor.
Installation instructions:
Deploying the functions as a MATLAB Runtime application for Mac with MATLAB
Compiler SDK has been a consistent issue and that version (which was the
best after many tries) only works on about 50% of Mac systems. I have
contacted Mathworks repeatedly but they don't have great support for Mac,
their only suggestion was to zip the application. Perhaps this should be a
Windows only release. Mac users that have MATLAB can run it in MATLAB
without having to use the MATLAB Runtime version.

@Kevin-Mattheus-Moerman <https://github.com/Kevin-Mattheus-Moerman> what's
the next step?

Aaron
