### 2018-12-12 #1094 @Kevin-Mattheus-Moerman
@carterenyi about the version issue. Yes you can create a release and let us know about the version number after the review process is complete.
Rebranding this as a Windows only release sounds like a good plan for now. Perhaps make changes to the readme/documentation to reflect this. Are you able to continue the review process on a Windows system? Thanks.
@ybayle thanks for your review efforts so far! Would you be able to focus on evaluation for Windows? Thanks.
