### 2018-12-19 carterenyi/emviz#4 @Kevin-Mattheus-Moerman
Add DOI's to paper references
Please check your references and add DOI's whenever they are available. Thanks.
