### 2018-12-19 #1094 @Kevin-Mattheus-Moerman
@ybayle thanks for these comments.

In relation to DOI's I've opened an issue on this (https://github.com/carterenyi/emviz/issues/4) whenever DOI's are available they need to be added. @carterenyi can you work on adding those?
