### 2018-12-19 #1094 @carterenyi
@Kevin-Mattheus-Moerman Yes, I will work on adding the DOIs

On Wed, Dec 19, 2018 at 5:30 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @ybayle <https://github.com/ybayle> thanks for these comments.
>
> In relation to DOI's I've opened an issue on this (carterenyi/emviz#4
> <https://github.com/carterenyi/emviz/issues/4>) whenever DOI's are
> available they need to be added. @carterenyi
> <https://github.com/carterenyi> can you work on adding those?
>
