### 2019-01-04 #1094 @Kevin-Mattheus-Moerman
Happy 2019 everybody.

@carterenyi how are you getting on? Did you manage to check those DOI's in the paper?

@ybayle, @gwynethbradbury can you provide an update on the review process? What are the major issues @carterenyi should work on? Thanks.
