### 2019-01-04 #1094 @ybayle
> Happy 2019 everybody.

Thank you and best wishes to you too.

> @ybayle can you provide an update on the review process?

There are still comments to be processed by the author from here: https://github.com/openjournals/joss-reviews/issues/1094#issuecomment-448377467

> What are the major issues @carterenyi should work on? Thanks.

The GUI does not display properly on Windows and the Github cloned version only works once.
