### 2019-01-05 #1094 @carterenyi
Hello All,

Thanks for checking in. I am in Nigeria without a strong internet
connection so it is not easy for me to work on the article right now. I
hope to complete the DOIs, GUI display and cloning issues by the end of the
month.

Aaron
