### 2019-01-31 #1094 @carterenyi
Hi Kevin,

Thanks so much for following up. Just got back to US and resumed teaching,
hope to complete the revisions by next week.

Aaron

On Thu, Jan 31, 2019 at 9:15 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @carterenyi <https://github.com/carterenyi> how are you getting on? Have
> you been able to resume work on this?
>
