### 2019-02-21 #1094 @carterenyi
Hi Kevin,

I've updated my matlab to 2018b on my mac and am now working on the
revisions. I've ordered a pc so that I can deploy the windows version. I
hope to have an update within a week.

On Sun, Feb 10, 2019 at 3:35 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @carterenyi <https://github.com/carterenyi> any updates on the review
> process? Thanks!
>
