+++ paper.bib
@@ -5,7 +5,8 @@
    volume = {22},
    number = {1},
    type = {Journal Article},
-   year = 2016
+   year = 2016,
+   url = {http://www.mtosmt.org/issues/mto.16.22.1/mto.16.22.1.carter-enyi.pdf}
 }
 
 @phdthesis{Carter-Enyi:2016b,
@@ -13,7 +14,8 @@
    title = {{Contour Levels: An Abstraction of Pitch Space based on African Tone Systems}},
    university = {Ohio State University},
    type = {Thesis},
-   year = 2016
+   year = 2016,
+   adsurl = {http://adsabs.harvard.edu/abs/2016PhDT.......160C}
 }
 
 @book{Huron:2006,
@@ -21,6 +23,7 @@
    title = {{Sweet Anticipation: Music and the Psychology of Expectation}},
    publisher = {MIT Press},
    address = {Cambridge},
+   ISBN = {9780262275965},
    type = {Book},
    year = 2006
 }
@@ -56,6 +59,7 @@
    number = {6},
    pages = {1596-1605},
    ISSN = {0001-4966},
+   DOI = {10.1121/1.387814}
    type = {Journal Article},
    year = 1982
 }
@@ -67,6 +71,7 @@
    publisher = {IEEE},
    pages = {110-116},
    ISBN = {076951751X},
+   DOI = {10.1109/INFVIS.2002.1173155},
    type = {Conference Proceedings},
    year = 2002
 }
