+++ paper.bib
@@ -15,7 +15,8 @@
    university = {Ohio State University},
    type = {Thesis},
    year = 2016,
-   adsurl = {http://adsabs.harvard.edu/abs/2016PhDT.......160C}
+   adsurl = {http://adsabs.harvard.edu/abs/2016PhDT.......160C},
+   url = {http://rave.ohiolink.edu/etdc/view?acc_num=osu1461029477}
 }
 
 @book{Huron:2006,
