### 2019-02-26 carterenyi/emviz#4 @carterenyi
I've added DOIs and where there are not DOIs, I added urls and/or ISBN
