+++ EMVizScripts/nmat2snd.m
@@ -1,257 +1,0 @@
-function w = nmat2snd(nmat,synthtype,fs)
-% Create waveform of NMAT using a simple synthesis
-% w = nmat2snd(nmat,<synthtype>,<fs>,<f_min>,<f_max>)
-% Create waveform of NMAT using a simple FM synthesis. The default sampling rate is 
-% 22050 Hz and velocities are scaled to have
-% a max value of 1.
-%
-% SYNTHTYPE 'fm' (default) uses FM synthesis to approximate horn sound.
-% SYNTHTYPE 'shepard' creates waveform of NMAT using Shepard tones. 
-% These tones have also been 
-%  called 'Circular tones' because they are specifically constructed to contain 
-%  frequency components at octave intervals with an emphasis of the spectral 
-%  components between 500Hz and 1000 Hz that effectively 
-%  eliminates octave information (Shepard, 1964). 
-%
-% Part of the code has been obtained from the work of Ed Doering.
-%  Ed.Doering@Rose-Hulman.Edu
-%
-% Input argument:
-%	NMAT = notematrix
-%     SYNTHTYPE (Optional) = Synthesis type, either FM synthesis ('fm', default) 
-%           or Shepard tones ('shepard')
-%     FS (optional) = sampling rate (default 22050)
-%
-% Output:
-%	Y = waveform
-%
-% Example 1: samples1 = nmat2snd(laksin);
-% Example 2: samples2 = nmat2snd(laksin,'shepard', 22050);
-%
-% Reference:
-%    Moore, F. R. (1990). Elements of Computer Music. New York: Prentice-Hall.
-%    Shepard, R. N. (1964). Circularity in judgements of 
-%       relative pitch. Journal of the Acoustical Society of America, 
-%       36, 2346-2353. 
-%
-% Change History :
-% Date		Time	Prog	Note
-% 29.9.2003	20:13	TE	Created under MATLAB 5.3 (PC)
-% 4/15/05           JBP added ability to specify number of partials in
-%                       tone, also shifted harmonic envelope such that tones are actually 3
-%                       semitones lower than desired. To fix, transpose input up by 3 semitones
-%                       (3 midi notes) to get correct pitch with better envelope.
-% 2.5.2005	9:30	TE	Almost a total revision if the shepard function
-%� Part of the MIDI Toolbox, Copyright � 2004, University of Jyvaskyla, Finland
-% See License.txt
-
-if isempty(nmat), return; end
-if nargin<3, fs = 22050; end
-if nargin<2, synthtype='fm'; end
-
-%%%%%%%%%%%%  WARNINGS
-sample_in_K = (fs*(1+nmat(end,6) + nmat(end,7)))/1024;
-
-if sample_in_K<1000
-	disp([num2str(sample_in_K,4) 'K samples']);
-elseif sample_in_K>1000
-    %TODO: this needs to be revised
-f=1;
-   while f==1
-	i = input([num2str(sample_in_K,4),'Ks of samples, synthesis may last for an extended period of time. \nDo you wish to continue? Y/N [Y]: '],'s');
-		if isempty(i)
-		    i = 'Y';
-		    f=2;
-		end
-	switch lower(i)
-	   case 'y'
-		    f=2;
-	   case 'n'
-		disp('synthesis cancelled')
-		y=[];	f=2; return
-	  otherwise
-	   	disp([i,' is not a valid choice'])
-		y=[];	
-		f=1;
-	end
-    end
-end	   
-
-% END WARNINGS
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-
-switch lower(synthtype)
-          case 'fm'
-
[…]
