### 2019-02-26 #1094 @carterenyi
On Sun, Feb 24, 2019 at 7:47 PM Aaron Carter-Enyi <carterenyi@gmail.com>
wrote:

> @Kevin-Mattheus-Moerman DOIs, urls and/or ISBNs have been added to all
> references where applicable. The windows laptop I ordered arrived yesterday
> so I should be able to fix the GUI display issue for Windows 10 this week.
>
> On Thu, Feb 21, 2019 at 2:41 PM Aaron Carter-Enyi <carterenyi@gmail.com>
> wrote:
>
>> Hi Kevin,
>>
>> I've updated my matlab to 2018b on my mac and am now working on the
>> revisions. I've ordered a pc so that I can deploy the windows version. I
>> hope to have an update within a week.
>>
>> On Sun, Feb 10, 2019 at 3:35 AM Kevin Mattheus Moerman <
>> notifications@github.com> wrote:
>>
>>> @carterenyi <https://github.com/carterenyi> any updates on the review
>>> process? Thanks!
>>>
>>> —
>>> You are receiving this because you were mentioned.
>>> Reply to this email directly, view it on GitHub
>>> <https://github.com/openjournals/joss-reviews/issues/1094#issuecomment-462114035>,
>>> or mute the thread
>>> <https://github.com/notifications/unsubscribe-auth/AXim6gwCZUYL8Kd3yGcqCTap3eQvA9U0ks5vL9nfgaJpZM4YtePz>
>>> .
>>>
>>
>>
>> --
>> Aaron Carter-Ényì, PhD, CAGO
>>
>
>
> --
> Aaron Carter-Ényì, PhD, CAGO
>

--
Aaron Carter-Ényì, PhD, CAGO
