+++ paper.bib
@@ -6,7 +6,7 @@
    number = {1},
    type = {Journal Article},
    year = 2016,
-   url = {http://www.mtosmt.org/issues/mto.16.22.1/mto.16.22.1.carter-enyi.pdf}
+   DOI = {doi.org/10.30535/mto.22.1.1}
 }
 
 @phdthesis{Carter-Enyi:2016b,
@@ -25,6 +25,7 @@
    publisher = {MIT Press},
    address = {Cambridge},
    ISBN = {9780262275965},
+   DOI = {10.7551/mitpress/6575.001.0001},
    type = {Book},
    year = 2006
 }
@@ -34,6 +35,7 @@
    title = {{Composition with pitch-classes: a theory of compositional design}},
    publisher = {Yale University Press},
    ISBN = {0300036841},
+   DOI = {10.2307/j.ctt1xp3ss4},
    type = {Book},
    year = 1987
 }
