+++ paper.bib
@@ -62,7 +62,7 @@
    number = {6},
    pages = {1596-1605},
    ISSN = {0001-4966},
-   DOI = {10.1121/1.387814}
+   DOI = {10.1121/1.387814},
    type = {Journal Article},
    year = 1982
 }
