+++ paper.bib
@@ -6,7 +6,7 @@
    number = {1},
    type = {Journal Article},
    year = 2016,
-   DOI = {doi.org/10.30535/mto.22.1.1}
+   DOI = {10.30535/mto.22.1.1}
 }
 
 @phdthesis{Carter-Enyi:2016b,
