### 2019-02-26 #1094 @carterenyi
@Kevin-Mattheus-Moerman, @ybayle, I've redeployed the software with MATLAB 2018b on Windows 10 and tested it. I've also added functionality testing instructions in the readme, generated documentation with m2html and updated information about windows only release in the paper and pdf. I did a v1.0 release on GitHub, but not sure if I did that right.
