### 2019-03-20 #1094 @ybayle
The revised version works fine for me.

`Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?`
This is the last check remaining. There are no automated tests implemented currently for the 3 parts of the algo (namely Data Preparation, Pattern Recognition and Visualization). How critical is this assertion for the review process?

Minor typos in the pdf:

- There are extra parenthesis for some citations  (e.g. `([@Wattenberg:2002])` instead of `[@Wattenberg:2002]`)
- `currently being developing Python`
