### 2019-03-24 #1094 @carterenyi
@Kevin-Mattheus-Moerman, @ybayle,

Manual steps are provided to verify functionality of the Pattern
Recognition and Visualization (with the csv output and arc diagram),
neither of which will be correct without the data preparation working. Is
that satisfactory?

I'll fix the minor typos.
