### 2019-04-07 #1094 @carterenyi
@Kevin-Mattheus-Moerman, @ybayle,

I've fixed the typos, are we good to go?
