### 2019-04-08 #1094 @Kevin-Mattheus-Moerman
> Manual steps are provided to verify functionality of the Pattern Recognition and Visualization (with the csv output and arc diagram), neither of which will be correct without the data preparation working. Is that satisfactory?

@carterenyi would it be easy to add a basic script to automate these manual steps?
