### 2019-04-08 #1094 @carterenyi
On Mon, Apr 8, 2019 at 4:46 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @carterenyi <https://github.com/carterenyi> some comments on the paper:
>
>    - "is brilliant" can you rephrase this more professionally? e.g.
>    perhaps you mean to say something like: innovative and useful?
>    - Missing ) bracket in: "(now available in MATLAB Runtime[)], and
>    ATAVizM (currently being developed in Python)"
>    - Should "C+" be written as "C++"?
>    - Are there any references/links (Zenodo, Github, website,...) for
>    "ATAVizM" since there are currently none (and I cannot find it through a
>    quick Google), should the comparison to "ATAVizM" (which is listed as in
>    development) perhaps be removed? Instead, perhaps mention it as future work
>    at the end?
>    - Table 1 lists "Add ..." for the ATAVizM components. Does this mean
>    these features are to be added? So this is a work in progress? Again
>    perhaps remove all references to this package here.
>    - Consider removing "—instead of just a pretty picture.", to me the
>    nice summary given is enough and this statement feels out of place.
>
