### 2019-04-08 #1094 @carterenyi
Hi @Kevin-Mattheus-Moerman,

Thanks for the suggested revisions and all your help through this process.

I have responded to all:
-revised wording as you suggested
-the check is basically is automated, the short CSV file and visualization
are the output of running the test file and the user can just briefly
compare it to those printed in the manual, I'm not sure how I would make it
more automated
-I did not remove ATAVizM, because that is the funded project of which this
is a basic version, I instead added some elaboration and links to the
website and a GitHub repository where those materials will be available in
the near future.
