### 2019-05-01 #1094 @ybayle
@carterenyi The provided test works for me.

However, this is not an automated test nor a suite of unit tests.
If I refer to this: `Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?`, yes there is one manual test that can check the output of the whole program but no there are not test verfying the output of each function.
@Kevin-Mattheus-Moerman How do you handle this? If you think this is enough, I can check the box, otherwise  tell me if I need to re-check something!
