### 2019-05-01 #1094 @carterenyi
Thanks @ybayle and @Kevin-Mattheus-Moerman for your patience with me. While
I know signal processing, I don't know much about software development
standards--and I'm learning a lot with this process!

Unfortunately, while I am understanding better the nature of an automated
test, it is not possible for me to redeploy the windows version of the
software right now (I bought a used windows machine but it stopped working
and I returned it). Is it possible to go with the manual tests?

Thanks,
Aaron
