### 2019-05-15 #1094 @Kevin-Mattheus-Moerman
@carterenyi at this point can you:

- [ ] Please archive a copy of the reviewed software on [Zenodo](http://about.zenodo.org/) (some find the steps defining [this](https://guides.github.com/activities/citable-code/) automated process useful). Once you've done that can you provide us with the DOI of the archived version?
- [ ] Can you please list the final version label for the reviewed software?
