### 2019-05-16 #1094 @carterenyi
@Kevin-Mattheus-Moerman Zenodo is down, but I will take care of that as
soon as it is up.

What do you mean by final version label (is it e.g. "1.0") and where should
I list it?

Thanks,
Aaron

On Wed, May 15, 2019 at 4:14 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @carterenyi <https://github.com/carterenyi> at this point can you:
>
>    - Please archive a copy of the reviewed software on Zenodo
>    <http://about.zenodo.org/> (some find the steps defining this
>    <https://guides.github.com/activities/citable-code/> automated process
>    useful). Once you've done that can you provide us with the DOI of the
>    archived version?
>    - Can you please list the final version label for the reviewed
>    software?
>
