### 2019-05-18 #1094 @carterenyi
@Kevin-Mattheus-Moerman I have taken care of the Zenodo DOI and listed the Version Release
