### 2019-05-18 #1094 @carterenyi
> @carterenyi at this point can you:
>
> * [x]  Please archive a copy of the reviewed software on [Zenodo](http://about.zenodo.org/) (some find the steps defining [this](https://guides.github.com/activities/citable-code/) automated process useful). Once you've done that can you provide us with the DOI of the archived version?
> * [x]  Can you please list the final version label for the reviewed software?
