### 2019-05-18 #1094 @Kevin-Mattheus-Moerman
@openjournals/joss-eics this submission is ready to be accepted. Can you confirm the title of this archived version is okay or does it need to be amended to match this submission? https://zenodo.org/record/2919897
