### 2019-05-18 #1094 @labarba
We prefer that the Zenodo deposit match the title of the paper. Also, note that here the author list needs editing, too. On editing the Zenodo metadata, I suggest that the author also add their ORCID.
