### 2019-05-18 #1094 @carterenyi
@openjournals/joss-eics
<https://github.com/orgs/openjournals/teams/joss-eics> I've made the
changes to Zenodo deposit and metadata

On Sat, May 18, 2019 at 1:27 PM Lorena A. Barba <notifications@github.com>
wrote:

> We prefer that the Zenodo deposit match the title of the paper. Also, note
> that here the author list needs editing, too. On editing the Zenodo
> metadata, I suggest that the author also add their ORCID.
>
