### 2019-05-18 #1094 @arfon
> Is @arfon a co-author??

Looks like I fixed the paper for the author: https://github.com/carterenyi/emviz/commits?author=arfon

@carterenyi - you should probably remove me from the authors on Zenodo :-)

Also, @labarba - please could I process the acceptance of this paper? I've been making some changes to Whedon I'd like to test when accepting this paper.
