### 2019-05-20 #1094 @carterenyi
Hello @labarba <https://github.com/labarba> and @arfon
<https://github.com/arfon>,

The only way I could figure out how to remove Arfon is to submit a new
version (v1.0.1) to zenodo.

Here is the new Zenodo DOI:
10.5281/zenodo.3051781 <https://doi.org/10.5281/zenodo.3051781>

Thank you @Kevin-Mattheus-Moerman and @ybayle <https://github.com/ybayle> for
all your time on this submission!

On Sat, May 18, 2019 at 3:09 PM Arfon Smith <notifications@github.com>
wrote:

> Is @arfon <https://github.com/arfon> a co-author??
>
> Looks like I fixed the paper for the author:
> https://github.com/carterenyi/emviz/commits?author=arfon
>
> @carterenyi <https://github.com/carterenyi> - you should probably remove
> me from the authors on Zenodo :-)
>
> Also, @labarba <https://github.com/labarba> - please could I process the
> acceptance of this paper? I've been making some changes to Whedon I'd like
> to test when accepting this paper.
>
