### 2019-05-20 #1094 @Kevin-Mattheus-Moerman
@arfon did you update the version tag to `v1.0.1`?
